refactored event kit code
